Update jquery.fb.ext.recaptcha-v3.js
Submitting a form with V3 using AJAX works if all fields are filled in. If any field fails validation and then the form resubmitted - the V3 captcha will fail. Logging the response from reCaptcha it says there failure reason is 'duplicate'. I believe this means the token sent has already been validated or attempted validation.

I propose resetting the token on any of the available events. This ensures a valid token is being sent for validation.